backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
python-multipart
python-socketio
python-jose[cryptography]
cloudinary
